make users who have verified email only who can create, edit and delete posts

+++ app/Http/Controllers/PostController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Enums\PostStatus;
+use Illuminate\Http\Request;
 use App\Http\Requests\PostRequest;
 use App\Models\{Category, User, Post};
 use Illuminate\Support\Facades\Storage;
@@ -26,8 +27,13 @@ class PostController extends Controller
         return view('posts.show', compact('post'));
     }
 
-    public function create()
+    public function create(Request $request)
     {
+        if ($request->user()->hasVerifiedEmail() == false) {
+            $request->user()->sendEmailVerificationNotification();
+            return to_route('verification.notice');
+        }
+
         return view('posts.create', [
             'post' => new Post(),
             'categories' => Category::all(),
@@ -36,12 +42,22 @@ class PostController extends Controller
 
     public function store(PostRequest $request)
     {
+        if ($request->user()->hasVerifiedEmail() == false) {
+            $request->user()->sendEmailVerificationNotification();
+            return to_route('verification.notice');
+        }
+
         $request->insert();
         return to_route('user_posts.index')->with('message', 'Post kamu berhasil dibuat :)');
     }
 
-    public function edit(Post $post)
+    public function edit(Request $request, Post $post)
     {
+        if ($request->user()->hasVerifiedEmail() == false) {
+            $request->user()->sendEmailVerificationNotification();
+            return to_route('verification.notice');
+        }
+
         return view('posts.edit', [
             'post' => $post,
             'categories' => Category::all(),
@@ -50,12 +66,22 @@ class PostController extends Controller
 
     public function update(PostRequest $request, Post $post)
     {
+        if ($request->user()->hasVerifiedEmail() == false) {
+            $request->user()->sendEmailVerificationNotification();
+            return to_route('verification.notice');
+        }
+
         $request->update($post);
         return to_route('user_posts.index')->with('message', 'Post kamu berhasil diubah :)');
     }
 
-    public function destroy(Post $post)
+    public function destroy(Request $request, Post $post)
     {
+        if ($request->user()->hasVerifiedEmail() == false) {
+            $request->user()->sendEmailVerificationNotification();
+            return to_route('verification.notice');
+        }
+
         !$post->image ?: Storage::delete($post->image);
         Post::destroy($post->id);
         return to_route('user_posts.index')->with('message', 'Post kamu berhasil dihapus :)');
